+++ src/argononed.h
@@ -89,7 +89,7 @@ struct SHM_Data {               //  DAEMON  |   CLIENT
     uint8_t req_flags;          //      RW  |   WO
     struct SHM_DAEMON_STATS stat;
     struct SHM_REQ_MSG msg; // Special Message for CLI client only
-    struct SHM_REQ_MSG msg_app[2]; // Normal Application Messages
+    struct SHM_REQ_MSG msg_app[2]; // Normal Application Messages **Not Yet Enabled**
 }; // current size - 14 bytes
 
 #endif